+++ test/clojure_lsp/db_test.clj
@@ -1,10 +1,10 @@
 (ns clojure-lsp.db-test
   (:require
    [clojure-lsp.db :as db]
+   [clojure-lsp.shared :as shared]
    [clojure-lsp.test-helper :as h]
    [clojure.java.io :as io]
-   [clojure.test :refer [deftest is testing]]
-   [medley.core :as medley]))
+   [clojure.test :refer [deftest is testing]]))
 
 (h/reset-db-after-test)
 
@@ -19,10 +19,10 @@
     (testing "when set to relative path"
       (let [settings-path "subdir/.cache"
             expected (.getAbsolutePath (io/file project-path settings-path))]
-        (swap! db/db medley/deep-merge {:settings {:cache-path settings-path}})
+        (swap! db/db shared/deep-merge {:settings {:cache-path settings-path}})
         (is (= expected
                (#'db/get-db-file-path project-path db/db)))))
     (testing "when set to absolute path"
-      (swap! db/db medley/deep-merge {:settings {:cache-path (h/file-path "/db-dir/.cache")}})
+      (swap! db/db shared/deep-merge {:settings {:cache-path (h/file-path "/db-dir/.cache")}})
       (is (= (h/file-path "/db-dir/.cache")
              (#'db/get-db-file-path project-path db/db))))))