+++ test/clojure_lsp/feature/clean_ns_test.clj
@@ -2,9 +2,9 @@
   (:require
    [clojure-lsp.db :as db]
    [clojure-lsp.feature.clean-ns :as f.clean-ns]
+   [clojure-lsp.shared :as shared]
    [clojure-lsp.test-helper :as h]
    [clojure.test :refer [deftest is testing]]
-   [medley.core :as medley]
    [rewrite-clj.zip :as z]))
 
 (h/reset-db-after-test)
@@ -16,7 +16,7 @@
    (test-clean-ns db input-code expected-code in-form "file:///a.clj"))
   ([db input-code expected-code in-form uri]
    (h/clean-db!)
-   (swap! db/db medley/deep-merge db)
+   (swap! db/db shared/deep-merge db)
    (h/load-code-and-locs input-code (h/file-uri uri))
    (let [zloc (when in-form
                 (-> (z/of-string input-code) z/down z/right z/right))