+++ test/clojure_lsp/settings_test.clj
@@ -2,18 +2,18 @@
   (:require
    [clojure-lsp.db :as db]
    [clojure-lsp.settings :as settings]
+   [clojure-lsp.shared :as shared]
    [clojure-lsp.test-helper :as h]
-   [clojure.test :refer [deftest is]]
-   [medley.core :as medley]))
+   [clojure.test :refer [deftest is]]))
 
 (h/reset-db-after-test)
 
 (deftest all-test
-  (swap! db/db medley/deep-merge {:settings {:a {:b {:c 2}}}})
+  (swap! db/db shared/deep-merge {:settings {:a {:b {:c 2}}}})
   (is (= {:a {:b {:c 2}}} (settings/all db/db))))
 
 (deftest get-test
-  (swap! db/db medley/deep-merge {:settings {:a {:b {:c 2}}}})
+  (swap! db/db shared/deep-merge {:settings {:a {:b {:c 2}}}})
   (is (= 2 (settings/get db/db [:a :b :c])))
   (is (= {:c 2} (settings/get db/db [:a :b])))
   (is (= {:b {:c 2}} (settings/get db/db [:a])))